Some code clean up

+++ db.class.php
@@ -6,7 +6,6 @@
 * (c) 22lbs Media - www.22lbs.ca
 *
 * This class is used to connect to a MySQL database and uses the MySQLi API introducted in PHP5. 
-* It has been inspired by WordPress' wpdb class
 *
 * Function examples are inline. If used properly you should never have to call a mysqli function directly in your application.
 * Activate by including this file in your php files and start the connection by creating a new class object. Example:
@@ -177,33 +176,48 @@ class DB {
 		$columns = "";
 		$values = "";
 
-		if ( is_array($data[0]) ) {
+		if ( is_array($data[0]) ) { // Check to see if multiple rows are to be inserted
+
 			foreach ( $data[0] as $column => $value ) {
+
 				$columns .= ($columns == "") ? "`" : "`,`";
 				$columns .= $column;
+
 			}
+
 			$columns .= "`";
 
+
 			foreach ( $data as $row ) {
+
 				foreach ( $row as $column => $value ) {
+
 					$values .= ($values == "") ? "('" : "'";
 					$values .= $this->escape($value);
 					$values .= "',";
+
 				}
+
 				$values = substr($values, 0, -1);
 				$values .= "), (";
+
 			}
+
 			$values = substr($values, 0, -3);
 
 			$sql = "INSERT INTO `" . $table . "` (" . $columns . ") VALUES " . $values;
 
 		} else {
+
 			foreach ($data as $column => $value) {
+
 				$columns .= ($columns == "") ? "`" : "`,`";
 				$columns .= $column;
 				$values .= ($values == "") ? "'" : "','";
 				$values .= $this->escape($value);
+
 			}
+
 			$columns .= "`";
 			$values .= "'";
 
@@ -212,20 +226,31 @@ class DB {
 		}
 		
 		if ( $this->query($sql) ) {
+
 			$this->last_id = $this->mysqli->insert_id;
 			return $this->last_id;
+
 		} else {
+
 			return false;
+
 		}
 		
 	}
 	
 	// update row
-	public function update($data=array(),$where='',$table='default_table') {
-	
-		$args=array();
-		foreach($data as $field=>$value){
-			$args[]="`".$field.'`="'.$this->escape($value).'"';
+	public function update($data=array(),$where=false,$table='default_table') {
+		
+		// Make sure that where is not empty
+		if ( $where === false ) {
+			return false;
+			die();
+		}
+
+		$args = array();
+
+		foreach( $data as $field => $value ) {
+			$args[] = "`" . $field . '` = "'. $this->escape($value) . '"';
 		}
 		
 		$sql = 'UPDATE ' . $table . ' SET '. implode(',',$args) . $this->where . $where;
@@ -235,8 +260,16 @@ class DB {
 	}
 	
 	// delete row(s)
-	public function delete($where='',$table='default_table'){
+	public function delete($where=false,$table='default_table'){
+
+		// Make sure that where is not empty
+		if ( $where === false ) {
+			return false;
+			die();
+		}
+
 		$sql=!$where?'DELETE FROM '.$table:'DELETE FROM '.$table.' WHERE '.$where;
-		$this->query($sql);
+		
+		return $this->query($sql);
 	}
 }
